requirements.txt
streamlit
streamlit-image-comparison
numpy
torch
matplotlib
pillow
opencv-python
scikit-image
torchvision
